requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
bcrypt==4.1.2
pyjwt==2.8.0
pydantic==2.7.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
# psycopg2==2.9.10